Added abstract PartDenseCable to reduce redundant code
